Clean-up and merged
