+++ map_gen/Diggy/Config.lua
@@ -1,85 +1,6 @@
 -- dependencies
+local MarketUnlockables = require 'map_gen.Diggy.Feature.MarketUnlockables'
 
-
-local marked_prototype_items = {}
-
-local function market_prototype_add(self_level, self_price, self_name)
-    if (not marked_prototype_items[self_level]) then
-        --market_prototype_items[''..self_level] = {}
-        table.insert(marked_prototype_items, self_level, {})
-    end
-    table.insert(marked_prototype_items[self_level], {price = self_price, name = self_name})
-end
-
-local function initalize_unlockables()
-  local levelcost = {}
-  local unlockables = {}
-  local prev_number = 0
-    for i = 1,100 do
-        local b = 20 -- Default 20 <-- Controls how much stone is needed.
-        local start_value = 50 -- The start value/the first level cost
-        local formula = b*(i^3)+(start_value-b)
-        
-        local precision = 2 -- Sets the precision
-        
-        --Truncates to the precision and prevents dublicates by incrementing with 5 in the third highest place
-        local number = formula
-        local numberlen = math.floor(math.log10(number)+1)
-        precision = (numberlen >= 8) and (precision+1) or precision
-        number = number/10^(numberlen-precision)
-        number = math.floor(number)*10^(numberlen-precision)
-        while (prev_number >= number) do
-            number = (prev_number > number) and number or number + math.ceil(5*10^(numberlen-3))
-        end
-        
-        --print(number .. " len "..numberlen .." DEBUG ".. tostring(numberlen >= 8) .. " precision ".. precision)
-        levelcost[i] = number
-        prev_number = number
-        
-        --[[local unlocking = {{name = 'mining_speed', value = 5}, {name = 'inventory_slot', value = 1}, {name = 'stone_automation', value = 3}}
-        for _, unlocked in ipairs(unlocking) do
-            local newTable = {stone = levelcost[i], type = 'buff',  prototype = unlocked}
-            table.insert(unlockables, newTable)
-        end]]
-        
-    end
-    -- market_prototype_add(unlock_level, price, prototype_name)
-    market_prototype_add(1, 50, 'raw-fish')
-    market_prototype_add(1, 50, 'steel-axe')
-    market_prototype_add(2, 50, 'small-electric-pole')
-    market_prototype_add(2, 50, 'small-lamp')
-    market_prototype_add(2, 25, 'stone-brick')
-    market_prototype_add(2, 125, 'stone-wall')
-    market_prototype_add(3, 850, 'submachine-gun')
-    market_prototype_add(3, 850, 'shotgun')
-    market_prototype_add(3, 50, 'firearm-magazine')
-    market_prototype_add(3, 50, 'shotgun-shell')
-    market_prototype_add(3, 500, 'light-armor')
-    market_prototype_add(11, 750, 'heavy-armor')
-    market_prototype_add(13, 100, 'piercing-rounds-magazine')
-    market_prototype_add(13, 100, 'piercing-shotgun-shell')
-    market_prototype_add(13, 1500, 'modular-armor')
-    market_prototype_add(16, 1000, 'landfill')
-    market_prototype_add(30, 250, 'uranium-rounds-magazine')
-    market_prototype_add(30, 1000, 'combat-shotgun')
-    
-    
-    
-
-      for lvl, v in pairs(marked_prototype_items) do
-          for _, w in ipairs(v) do
-              table.insert(unlockables, {level = lvl, stone = levelcost[lvl], type = 'market', prototype = w})
-          end
-      end
-      
-      --[[for _, v in ipairs(unlockables) do
-        print(v.stone)
-          for _, v in ipairs(v.prototype) do
-          end
-      end]]
-      
-    return unlockables
-end
 -- this
 local Config = {
     -- enable debug mode, shows extra messages
@@ -360,224 +281,12 @@ local Config = {
             -- every x ticks it will clear y currency_item
             void_chest_frequency = 307,
             
-            unlockables = initalize_unlockables(),
-            buffs = {
+            unlockables = MarketUnlockables.initalize_unlockables(), --Define new items in MarketUnlockables.lua
+            buffs = { --Define new buffs here
                 {prototype = {name = 'mining_speed', value = 5}},
                 {prototype = {name = 'inventory_slot', value = 1}},
                 {prototype = {name = 'stone_automation', value = 3}},
             }
-
-            --[[unlockables = {
-              
-                
-                {stone = 50, type = 'buff', prototype = {name = 'mining_speed', value = 10}},
-                {stone = 50, type = 'buff', prototype = {name = 'inventory_slot', value = 3}},
-                {stone = 50, type = 'buff', prototype = {name = 'stone_automation', value = 10}},
-                {stone = 50, type = 'market', prototype = {price = 25, name = 'raw-wood'}},
-                {stone = 50, type = 'market', prototype = {price = 50, name = 'raw-fish'}},
-                {stone = 50, type = 'market', prototype = {price = 175, name = 'steel-axe'}},
-
-                {stone = 250, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 250, type = 'buff', prototype = {name = 'inventory_slot', value = 2}},
-                {stone = 250, type = 'buff', prototype = {name = 'stone_automation', value = 5}},
-                {stone = 250, type = 'market', prototype = {price = 50, name = 'small-lamp'}},
-                {stone = 250, type = 'market', prototype = {price = 25, name = 'stone-brick'}},
-                {stone = 250, type = 'market', prototype = {price = 125, name = 'stone-wall'}},
-
-                {stone = 450, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 450, type = 'buff', prototype = {name = 'inventory_slot', value = 2}},
-                {stone = 450, type = 'buff', prototype = {name = 'stone_automation', value = 5}},
-                {stone = 450, type = 'market', prototype = {price = 850, name = 'submachine-gun'}},
-                {stone = 450, type = 'market', prototype = {price = 850, name = 'shotgun'}},
-                {stone = 450, type = 'market', prototype = {price = 50, name = 'firearm-magazine'}},
-                {stone = 450, type = 'market', prototype = {price = 50, name = 'shotgun-shell'}},
-                {stone = 450, type = 'market', prototype = {price = 500, name = 'light-armor'}},
-
-                {stone = 750, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 750, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 750, type = 'buff', prototype = {name = 'stone_automation', value = 5}},
-
-                {stone = 1250, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 1250, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 1250, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 1750, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 1750, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 1750, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 2500, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 2500, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 2500, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 4000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 4000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 4000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 6500, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 6500, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 6500, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 8000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 8000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 8000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 10000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 10000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 10000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 10000, type = 'market', prototype = {price = 750, name = 'heavy-armor'}},
-
-                {stone = 15000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 15000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 25000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 25000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 25000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 35000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 35000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 35000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 35000, type = 'market', prototype = {price = 100, name = 'piercing-rounds-magazine'}},
-                {stone = 35000, type = 'market', prototype = {price = 100, name = 'piercing-shotgun-shell'}},
-                {stone = 35000, type = 'market', prototype = {price = 1500, name = 'modular-armor'}},
-
-                {stone = 50000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 50000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 50000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 75000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 75000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 75000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 100000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 100000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 100000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 100000, type = 'market', prototype = {price = 1000, name = 'landfill'}},
-
-                {stone = 125000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 125000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 125000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 150000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 150000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 150000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 175000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 175000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 175000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 200000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 200000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 200000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 225000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 225000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 225000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 250000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 250000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 250000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 275000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 275000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 275000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 300000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 300000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 300000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 350000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 350000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 350000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 400000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 400000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 400000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 500000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 500000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 500000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 600000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 600000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 600000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 700000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 700000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 700000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-
-                {stone = 800000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 800000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-                {stone = 800000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 800000, type = 'market', prototype = {price = 250, name = 'uranium-rounds-magazine'}},
-                {stone = 800000, type = 'market', prototype = {price = 1000, name = 'combat-shotgun'}},
-
-                {stone = 900000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 900000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 900000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 1000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 1000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 1000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 1250000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 1250000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 1250000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 1500000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 1500000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 1500000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 1750000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 1750000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 1750000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 2000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 2000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 2000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 2500000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 2500000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 2500000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 3000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 3000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 3000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 3500000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 3500000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 3500000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 4000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 4000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 4000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 4500000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 4500000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 4500000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 5000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 5000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 5000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 6000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 6000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 6000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 7000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 7000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 7000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 8000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 8000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 8000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 9000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 9000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 9000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-
-                {stone = 10000000, type = 'buff', prototype = {name = 'mining_speed', value = 5}},
-                {stone = 10000000, type = 'buff', prototype = {name = 'stone_automation', value = 2}},
-                {stone = 10000000, type = 'buff', prototype = {name = 'inventory_slot', value = 1}},
-            },]]
         },
     },
 }
